feat: Trip 연관 엔티티 코드 정리
