+++ src/main/java/com/example/yeobee/core/currency/domain/Currency.java
@@ -1,10 +1,8 @@
 package com.example.yeobee.core.currency.domain;
 
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.Getter;
 
 @Entity
@@ -16,9 +14,6 @@ public class Currency {
 
     private String name;
 
-    @OneToMany(mappedBy = "currency")
-    private List<CountryCurrency> countryCurrencyList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "currency")
-    private List<TripCurrency> tripCurrencyList = new ArrayList<>();
+    @Embedded
+    private ExchangeRate exchangeRate;
 }
