+++ src/main/java/com/example/yeobee/core/currency/domain/ExchangeRate.java
@@ -1,8 +1,6 @@
 package com.example.yeobee.core.currency.domain;
 
 import jakarta.persistence.Embeddable;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import lombok.Getter;
@@ -10,9 +8,6 @@ import lombok.Getter;
 @Embeddable
 @Getter
 public class ExchangeRate {
-
-    @Enumerated(EnumType.STRING)
-    private ExchangeRateType exchangeRateType;
 
     private BigDecimal exchangeRateValue;
