--- src/main/java/com/example/yeobee/core/currency/domain/ExchangeRateType.java
@@ -1,7 +1,0 @@
-package com.example.yeobee.core.currency.domain;
-
-public enum ExchangeRateType {
-
-    CUSTOM,
-    AUTO
-}
